backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.14.0
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.2.0
blinker==1.9.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==41.0.1
discord==2.1.0
discord.py==2.1.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.6
MarkupSafe==2.1.1
MouseInfo==0.1.3
multidict==6.0.4
mysql-connector-python==9.1.0
PyAutoGUI==0.9.53
pycparser==2.21
pygame==2.1.2
PyGetWindow==0.0.9
PyJWT==2.9.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dotenv==1.0.1
pytweening==1.0.4
requests==2.28.1
rubicon-objc==0.4.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==1.26.12
Werkzeug==3.1.3
yarl==1.8.2
